UpdateChart removesChart unnecessary
If an ember model is changed, the chart data has to update as well. This is done by destroying the chart + building a new one. I can see why this is done, as the structure of the data can change as well (by adding more or less data items).

But if the structure does not change, it's better to only update the values. All the items will be animated to the new value. Currently all the data is set to 0 and then the animation begins.