provide a wrapper for canvas since Chart.js hardly relies on parentNode
Chart.js adds a hidden iframe to parentNode in `helpers.addResizeListener`
method and calculates maximum width / height depending on styles of parent
node in `helpers.getMaximumWidth` and `getMaximumHeight` methods.

In Microsoft edge there are cases (propably connected to destroy) in which
`parentNode` is `null` and therefore chart.js throws an error
`Unable to get property 'currentStyle' of undefined or null` in `getStyle`
method called by `getMaximumWidth` and `getMaximumHeight`.