Change to didUpdateAttrs() from observers
Changed to use didUpdateAttrs instead of observers.
Not only is this the standard ember way now, this also prevent firing update() multiple times if multiple changes occur in a single render cycle.
Also added a animate property to allow the user to decide when to animte. 
This allows disabling animations on update renders if they don't want a reanimation (like a onClick selection of an element in my case).